add : poster back drop content component
